refresh dir
